requirements.txt
python-dotenv
langchain_anthropic
langchain_core
pydantic
langgraph
langchain
langchain_mcp_adapters
langchain-openai
langgraph-supervisor
ipython
langsmith